backend/app/requirements.txt
Flask==3.0.2
python-docx==1.1.0